setup startup as brain. add new startup intended for auto brain pull
